refactor: extract hooks into separate folder
