+++ src/App.css
@@ -1,0 +1,5 @@
+.bird__image {
+  width: 150px;
+  height: 150px;
+  object-fit: cover;
+}
