--- src/components/info/Info.css
@@ -1,5 +1,0 @@
-.info__image {
-  width: 150px;
-  height: 150px;
-  object-fit: cover;
-}
